start manage exceptions

+++ services/activitiesService.ts
@@ -1,4 +1,5 @@
 import { IActivitiesModel, ActivitiesModel } from "../models/activitiesModel";
+import ValidationError from "../utils/ValidationError";
 
 
 async function getAllActivities():Promise<IActivitiesModel[]>{
@@ -10,16 +11,26 @@ async function getOneActivitie( _id: string ):Promise<IActivitiesModel[]>{
 }
 
 async function saveOneActivitie( activitie:IActivitiesModel ):Promise<IActivitiesModel>{
+    const err = activitie.validateSync();
+    if (err)
+        throw new ValidationError(err.message);
     const newActivitie = await activitie.save();
     return newActivitie;
 }
 
 async function updateOneActivitie( _id: string, activitie:IActivitiesModel ):Promise<IActivitiesModel>{
+    const err = activitie.validateSync();
+    if (err)
+        throw new ValidationError(err.message);
     const newActivitie = await ActivitiesModel.findByIdAndUpdate( _id, activitie, { returnOriginal: false } );
     return newActivitie;
 }
 
 async function deleteOneActivitie( _id: string ):Promise<void>{
+    const activity = await getOneActivitie(_id);
+    if (!activity)
+        throw new ValidationError("activity not found");
+
     return await ActivitiesModel.findByIdAndDelete( _id );
 }
